requirements.txt
flask
dotenv
os
openai
langchain
requests
json
newsapi
datetime
google-generativeai